requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements-dev.txt requirements-dev.in
#
asgiref==3.7.2
    # via django
asttokens==2.4.1
    # via stack-data
black==23.12.1
    # via -r requirements-dev.in
build==1.0.3
    # via pip-tools
cachetools==5.3.2
    # via tox
certifi==2023.11.17
    # via requests
chardet==5.2.0
    # via tox
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via tox
coverage==7.4.0
    # via -r requirements-dev.in
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via django-bootstrap-icons
distlib==0.3.8
    # via virtualenv
django==4.2.9
    # via
    #   -r requirements.in
    #   django-bootstrap-icons
    #   django-debug-toolbar
    #   django-distill
    #   django-extensions
    #   django-genericsite
    #   django-rich
    #   django-taggit
    #   easy-thumbnails
django-bootstrap-icons==0.8.7
    # via django-genericsite
django-debug-toolbar==4.2.0
    # via -r requirements-dev.in
django-distill @ https://github.com/veselosky/django-distill/releases/download/v3.1.3_vv01/django_distill-3.1.3-py3-none-any.whl
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-genericsite @ https://github.com/veselosky/django-genericsite/releases/download/release-0.1.7/django_genericsite-0.1.7-py3-none-any.whl
    # via -r requirements.in
django-rich==1.8.0
    # via -r requirements.in
django-taggit==5.0.1
    # via django-genericsite
django-tinymce==3.6.1
    # via django-genericsite
docutils==0.20.1
    # via
    #   -r requirements.in
    #   django-genericsite
easy-thumbnails==2.8.5
    # via django-genericsite
exceptiongroup==1.2.0
    # via ipython
executing==2.0.1
    # via stack-data
filelock==3.13.1
    # via
    #   tox
    #   virtualenv
gunicorn==21.2.0
    # via -r requirements.in
idna==3.6
    # via requests
ipython==8.19.0
    # via -r requirements-dev.in
isort==5.13.2
    # via -r requirements-dev.in
jedi==0.19.1
    # via ipython
markdown-it-py==3.0.0
    # via rich
matplotlib-inline==0.1.6
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   black
    #   build
    #   gunicorn
    #   pyproject-api
    #   tox
parso==0.8.3
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pillow==10.2.0
    # via
    #   -r requirements.in
    #   django-genericsite
    #   easy-thumbnails
pip-tools==7.3.0
    # via -r requirements-dev.in
platformdirs==4.1.0
    # via
    #   black
    #   tox
    #   virtualenv
pluggy==1.3.0
    # via tox
prompt-toolkit==3.0.43
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pydantic==1.10.13
    # via django-genericsite
pygments==2.17.2
    # via
    #   ipython
    #   rich
pyproject-api==1.6.1
    # via tox
pyproject-hooks==1.0.0
    # via build
requests==2.31.0
    # via
    #   django-bootstrap-icons
    #   django-distill
rich==13.7.0
    # via
    #   -r requirements.in
    #   django-rich
setproctitle==1.3.3
    # via -r requirements.in
six==1.16.0
    # via asttokens
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-api
    #   pyproject-hooks
    #   tox
tox==4.11.4
    # via -r requirements-dev.in
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.9.0
    # via
    #   asgiref
    #   black
    #   pydantic
urllib3==2.1.0
    # via requests
virtualenv==20.25.0
    # via tox
wcwidth==0.2.12
    # via prompt-toolkit
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
pip==23.3.2
    # via pip-tools
setuptools==69.0.3
    # via pip-tools
